Add moving-backward-page and moving-forward-page
Added to "replace" backward-page and forward-page, which don't play quite as
well as they could with page-break-lines-mode, as the cursor appears to end
up at the end of the line rather than the beginning. So this does the page
back/forward and then ensures we're at the start of the line.